add auth headers to workout requests

+++ src/components/workout/Workout.js
@@ -1,12 +1,19 @@
 import { useWorkoutsContext } from "../../hooks/useWorkoutsContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../../hooks/useAuthContext";
 
 const Workout = ({ workout }) => {
     const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
 
     const handleDelete = async () => {
+        if (!user) return;
+
         const response = await fetch("/api/workouts/" + workout._id, {
             method: "DELETE",
+            headers: {
+                Authorization: `Bearer ${user.token}`,
+            },
         });
         const json = await response.json();
